// api/api.ts
export const AuthMethod = {
    Password: 'Password',
    Email: 'Email',
    Mobile: 'Mobile'
} as const;

export type AuthMethod = typeof AuthMethod[keyof typeof AuthMethod];


/**
 * 
 * @export
 * @interface AuthorizeCommand
 */
export interface AuthorizeCommand {
    /**
     * 
     * @type {string}
     * @memberof AuthorizeCommand
     */
    'callbackURL'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuthorizeCommand
     */
    'clientId'?: string;
}
// ...
export interface LogoutCommand {
    /**
     * 
     * @type {string}
     * @memberof LogoutCommand
     */
    'logoutCallbackUrl'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LogoutCommand
     */
    'notificationsSubId'?: string;
}
// ...
export interface PasswordLoginCommand {
    /**
     * 
     * @type {string}
     * @memberof PasswordLoginCommand
     */
    'email'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PasswordLoginCommand
     */
    'password'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PasswordLoginCommand
     */
    'requestKey'?: string | null;
}
// ...
export interface RedirectDto {
    /**
     * 
     * @type {string}
     * @memberof RedirectDto
     */
    'redirectUrl'?: string | null;
}
/**
 * 
 * @export
 * @interface RegisterServiceManagerCommand
 */
export interface RegisterServiceManagerCommand {
    /**
     * 
     * @type {string}
     * @memberof RegisterServiceManagerCommand
     */
    'firstname'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RegisterServiceManagerCommand
     */
    'lastname'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RegisterServiceManagerCommand
     */
    'email'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RegisterServiceManagerCommand
     */
    'organization'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RegisterServiceManagerCommand
     */
    'password'?: string | null;
}
// ...
export interface TokenCommand {
    /**
     * 
     * @type {string}
     * @memberof TokenCommand
     */
    'clientId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TokenCommand
     */
    'authToken'?: string | null;
}
// ...
export interface VerifyLogoutCommand {
    /**
     * 
     * @type {string}
     * @memberof VerifyLogoutCommand
     */
    'logoutToken'?: string | null;
}
// ...
export class AuthApi extends BaseAPI {
    /**
     * 
     * @param {AuthorizeCommand} [authorizeCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authAuthorize(authorizeCommand?: AuthorizeCommand, options?: AxiosRequestConfig) {
        return AuthApiFp(this.configuration).authAuthorize(authorizeCommand, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authCheck(options?: AxiosRequestConfig) {
        return AuthApiFp(this.configuration).authCheck(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} [requestKey] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authGetLoginRequestDescription(requestKey?: string, options?: AxiosRequestConfig) {
        return AuthApiFp(this.configuration).authGetLoginRequestDescription(requestKey, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {LogoutCommand} [logoutCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authLogout(logoutCommand?: LogoutCommand, options?: AxiosRequestConfig) {
        return AuthApiFp(this.configuration).authLogout(logoutCommand, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {PasswordLoginCommand} [passwordLoginCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authPwdLogin(passwordLoginCommand?: PasswordLoginCommand, options?: AxiosRequestConfig) {
        return AuthApiFp(this.configuration).authPwdLogin(passwordLoginCommand, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authRefreshToken(options?: AxiosRequestConfig) {
        return AuthApiFp(this.configuration).authRefreshToken(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {RegisterServiceManagerCommand} [registerServiceManagerCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authRegisterManager(registerServiceManagerCommand?: RegisterServiceManagerCommand, options?: AxiosRequestConfig) {
        return AuthApiFp(this.configuration).authRegisterManager(registerServiceManagerCommand, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {TokenCommand} [tokenCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authToken(tokenCommand?: TokenCommand, options?: AxiosRequestConfig) {
        return AuthApiFp(this.configuration).authToken(tokenCommand, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {VerifyLogoutCommand} [verifyLogoutCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AuthApi
     */
    public authVerifyLogout(verifyLogoutCommand?: VerifyLogoutCommand, options?: AxiosRequestConfig) {
        return AuthApiFp(this.configuration).authVerifyLogout(verifyLogoutCommand, options).then((request) => request(this.axios, this.basePath));
    }
}